SimpleInteractiveObjects done, time for advanced now.
All of the current dependents of simpleInteractiveObjects now inherit from the abstract class and in extension, the interface above it. This satisfies the strategy design pattern.

Next step is to incorporate the hugely important player class into an AdvancedInteractiveObjects abstract class and trying adding new features like a different kind of powerup to check adherence to design patterns like OCP and DSP. The advantages of the singleton class system is already apparent with the members of the PlayerController Singleton Class being so readily and easily accessible all throughout the code.

We also need to perform an SRP analysis. The code is not the best/cleanest thing ever, but it is miles better than the spaghetti code it was before. Some explosions and redundant actions do need to be minimized though, but that will come right after the DestroyByContact and Mover classes are refactored completely.